working on homeroutes and added a test seed
We need to reorganized the routes. apiHomeroutes is the old homeroutes and it need to go in the the api folder but i have not move it so it does not affect other code
