+++ controllers/index.js
@@ -1,7 +1,12 @@
 const router = require('express').Router();
 
-const homeRoutes = require('./homeRoutes');
+const homeRoutes = require('./homeRoutes.js');
+//do not change!!! this still works!!!!!
+const apiHomeRoutes = require('./apiHomeRoutes.js');
+
 
 router.use('/', homeRoutes);
+router.use('/api', apiHomeRoutes);
+
 
 module.exports = router;
