+++ seeds/user.js
@@ -1,0 +1,1 @@
+//if we want to add a test users
